refactor: 의존관계 수정
